+++ cmd/culture.go
@@ -28,17 +28,21 @@ var cultureCmd = &cobra.Command{
 		random.SeedFromString(randomSeed)
 		randomCulture, err := culture.Random()
 		if err != nil {
-			fmt.Println(fmt.Errorf("Failed to generate culture: %w", err))
+			fmt.Println(fmt.Errorf("failed to generate culture: %w", err))
 			os.Exit(1)
 		}
 
 		res, err := json.Marshal(randomCulture)
 		if err != nil {
-			fmt.Println(fmt.Errorf("Failed to generate culture: %w", err))
+			fmt.Println(fmt.Errorf("failed to convert culture to json: %w", err))
 			os.Exit(1)
 		}
 		if outputFile != "" {
 			err = os.WriteFile(outputFile, res, 0644)
+			if err != nil {
+				fmt.Println(fmt.Errorf("failed to write culture to file: %w", err))
+				os.Exit(1)
+			}
 		} else {
 			fmt.Print(string(res))
 		}
@@ -47,14 +51,4 @@ var cultureCmd = &cobra.Command{
 
 func init() {
 	rootCmd.AddCommand(cultureCmd)
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// cultureCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// cultureCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }