+++ cmd/region.go
@@ -28,17 +28,21 @@ var regionCmd = &cobra.Command{
 		
 		result, err := region.Random()
 		if err != nil {
-			fmt.Println(fmt.Errorf("Failed to generate region: %w", err))
+			fmt.Println(fmt.Errorf("failed to generate region: %w", err))
 			os.Exit(1)
 		}
 
 		res, err := json.Marshal(result)
 		if err != nil {
-			fmt.Println(fmt.Errorf("Failed to generate region: %w", err))
+			fmt.Println(fmt.Errorf("failed to generate region: %w", err))
 			os.Exit(1)
 		}
 		if outputFile != "" {
 			err = os.WriteFile(outputFile, res, 0644)
+			if err != nil {
+				fmt.Println(fmt.Errorf("failed to write region to file: %w", err))
+				os.Exit(1)
+			}
 		} else {
 			fmt.Print(string(res))
 		}
@@ -47,14 +51,4 @@ var regionCmd = &cobra.Command{
 
 func init() {
 	rootCmd.AddCommand(regionCmd)
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// regionCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// regionCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }